Speed Changes
This allows the Ajax Request to process faster as I changed the code to
call less queries. I have made a number of other changes. I am going to
attempt to turn this into a module soon.

Some problems: The url redirection is primitive and doesn't work
without the presence of the layered nav block.
Obviously the code isn't very flexible and now can only be used for
very specific cases.